Moved DrawingPanelSimple into separate class, minor changes
